правки для #10 Шаблонизация проекта
