+++ functions.php
@@ -1,5 +1,14 @@
 <?php 
-function renderTemplate($path, $varArray) {
+/*
+* Возвращает результат сборки страницы.
+*
+* @param sting $path
+* @param array $varArray
+* @return string
+*/
+function toRenderTemplate($path, $varArray) {
+
+    $path = "templates/" . $path;
 
     if (!file_exists($path)) {
         return "";
@@ -11,4 +20,28 @@ function renderTemplate($path, $varArray) {
     
     return ob_get_clean(); 
 }
+
+/*
+* Возвращает время в относительном формате.
+*
+* @param number $time
+* @return string
+*/
+function convertUnixTime($time) {
+    
+        $oneDay = 86400;
+        $oneHour = 3600;
+    
+        $time = time() - $time; 
+    
+        if ($time >= $oneDay) {
+            return date('d.m.y \в H:i', $time);
+        }
+    
+        if ($time < $oneDay && $time >= $oneHour) {
+            return  date('h', $time) . ' часов назад';
+        }
+    
+        return date('i', $time) . ' минут назад';
+    }
 ?>
