+++ lot.php
@@ -1,4 +1,5 @@
 <?php
+require_once('functions.php');
 
 // ставки пользователей, которыми надо заполнить таблицу
 $bets = [
@@ -7,31 +8,6 @@ $bets = [
     ['name' => 'Евгений', 'price' => 10500, 'ts' => strtotime('-' . rand(25, 50) .' hour')],
     ['name' => 'Семён', 'price' => 10000, 'ts' => strtotime('last week')]
 ];
-
-/*
-* Возвращает время в относительном формате.
-*
-* @param {number} $time
-* @return {string}
-*/
-function convertUnixTime($time) {
-
-    $oneDay = 86400;
-    $oneHour = 3600;
-
-    $time = time() - $time; 
-
-    if ($time >= $oneDay) {
-        return date('d.m.y \в H:i', $time);
-    }
-
-    if ($time < $oneDay && $time >= $oneHour) {
-        return  date('h', $time) . ' часов назад';
-    }
-
-    return date('i', $time) . ' минут назад';
-}
-
 ?>
 
 <!DOCTYPE html>
